update naming, delete none
